// types.ts
export interface ContactInfo {
  name: string;
  email: string[] | null;
  linkedinUrl: string | null;
  phoneNumber: string[] | null;
  website: string | null;
  lineId: string | null;
  companyName: string | null;
  jobTitle: string | null;
  address: string | null;
  qrCodeUrl: string | null;
  otherInfo: string | null;
}